falta testar se as minas ta tudo certinho

+++ jeancp/src/circulo.h
@@ -1,10 +1,9 @@
 #ifndef _CIRCULO_H
 #define _CIRCULO_H
 
-#include <stdio.h>
-#include <stdlib.h>
 #include <string.h>
 #include "lista.h"
+#include "barco.h"
 
 /**
 CIRCULO é um polígono que possui um ponto de origem (âncora definida por x e y), raio, cor de borda e
@@ -12,7 +11,7 @@ cor de preenchimento.
 */
 typedef void *circulo;
 
-circulo create_circulo(int id, float x, float y, float r, char* corb, char* corp);
+circulo create_circulo(int id, double x, double y, double r, char* corb, char* corp);
 
 //CREATE_CIRCULO retorna um CIRCULO;
 
@@ -20,17 +19,17 @@ void circulo_set_i(circulo cir, int id);
 
 //CIRCULO_SET_I recebe um CIRCULO e um inteiro e seta o ID do CIRCULO
 
-void circulo_set_x(circulo cir, float x);
+void circulo_set_x(circulo cir, double x);
 
-//CIRCULO_SET_X recebe um CIRCULO e um float e seta o x da ancora (centro) do CIRCULO
+//CIRCULO_SET_X recebe um CIRCULO e um double e seta o x da ancora (centro) do CIRCULO
 
-void circulo_set_y(circulo cir, float y);
+void circulo_set_y(circulo cir, double y);
 
-//CIRCULO_SET_Y recebe um CIRCULO e um float e seta o y da ancora (centro) do CIRCULO
+//CIRCULO_SET_Y recebe um CIRCULO e um double e seta o y da ancora (centro) do CIRCULO
 
-void circulo_set_r(circulo cir, float r);
+void circulo_set_r(circulo cir, double r);
 
-//CIRCULO_SET_R recebe um CIRCULO e um float e seta o raio do CIRCULO
+//CIRCULO_SET_R recebe um CIRCULO e um double e seta o raio do CIRCULO
 
 void circulo_set_corb(circulo cir, char* corb);
 
@@ -44,15 +43,15 @@ int circulo_get_i(circulo cir);
 
 //CIRCULO_GET_I recebe um CIRCULO e retorna o ID do CIRCULO
 
-float circulo_get_x(circulo cir);
+double circulo_get_x(circulo cir);
 
 //CIRCULO_GET_X recebe um CIRCULO e retorna o x da ancora (centro) do CIRCULO
 
-float circulo_get_y(circulo cir);
+double circulo_get_y(circulo cir);
 
 //CIRCULO_GET_Y recebe um CIRCULO e retorna o y da ancora (centro) do CIRCULO
 
-float circulo_get_r(circulo cir);
+double circulo_get_r(circulo cir);
 
 //CIRCULO_GET_R recebe um CIRCULO e retorna o raio do CIRCULO
 
@@ -64,7 +63,7 @@ char* circulo_get_corp(circulo cir);
 
 //CIRCULO_GET_CORP recebe um CIRCULO e retorna a cor de preenchimento do CIRCULO
 
-float circulo_get_area(circulo cir);
+double circulo_get_area(circulo cir);
 
 //CIRCULO_GET_AREA recebe um CIRCULO e retorna a área do CIRCULO
