+++ jeancp/src/linha.c
@@ -4,16 +4,16 @@
 struct Line
 {
     int id;
-    float x1;
-    float y1;
-    float x2;
-    float y2;
+    double x1;
+    double y1;
+    double x2;
+    double y2;
     char* cor;
     int hp;
     int np;
 };
 
-linha create_linha(int id, float x1, float y1, float x2, float y2, char* cor)
+linha create_linha(int id, double x1, double y1, double x2, double y2, char* cor)
 {
     struct Line *li = malloc(sizeof(struct Line));
     linha_set_i(li, id);
@@ -31,25 +31,25 @@ void linha_set_i(linha li, int i)
     ponteiro->id = i;
 }
 
-void linha_set_x1(linha li, float x1)
+void linha_set_x1(linha li, double x1)
 {
     struct Line *ponteiro = li;
     ponteiro->x1 = x1;
 }
 
-void linha_set_y1(linha li, float y1)
+void linha_set_y1(linha li, double y1)
 {
     struct Line *ponteiro = li;
     ponteiro->y1 = y1;
 }
 
-void linha_set_x2(linha li, float x2)
+void linha_set_x2(linha li, double x2)
 {
     struct Line *ponteiro = li;
     ponteiro->x2 = x2;
 }
 
-void linha_set_y2(linha li, float y2)
+void linha_set_y2(linha li, double y2)
 {
     struct Line *ponteiro = li;
     ponteiro->y2 = y2;
@@ -69,25 +69,25 @@ int linha_get_i(linha li)
     return ponteiro->id;
 }
 
-float linha_get_x1(linha li)
+double linha_get_x1(linha li)
 {
     struct Line *ponteiro = li;
     return ponteiro->x1;
 }
 
-float linha_get_y1(linha li)
+double linha_get_y1(linha li)
 {
     struct Line *ponteiro = li;
     return ponteiro->y1;
 }
 
-float linha_get_x2(linha li)
+double linha_get_x2(linha li)
 {
     struct Line *ponteiro = li;
     return ponteiro->x2;
 }
 
-float linha_get_y2(linha li)
+double linha_get_y2(linha li)
 {
     struct Line *ponteiro = li;
     return ponteiro->y2;
