+++ jeancp/src/mina.h
@@ -1,8 +1,6 @@
 #ifndef _MINA_H_
 #define _MINA_H_
 
-#include <stdio.h>
-#include <stdlib.h>
 #include "texto.h"
 
 /**
@@ -11,16 +9,16 @@ Uma mina é um objeto que pode ser colocado no campo de batalha e que explode qu
 typedef void* Mina;
 
 /** Cria uma mina */
-Mina createMina(int id, float x, float y, char* corb, char* corp, char* ancora);
+Mina createMina(int id, double x, double y, char* corb, char* corp, char* ancora);
 
 /** Seta o id da mina */
 void setMinaId(Mina mina, int id);
 
 /** Seta a coordenada x da ancora da mina */
-void setMinaX(Mina mina, float x);
+void setMinaX(Mina mina, double x);
 
 /** Seta a coordenada y da ancora da mina */
-void setMinaY(Mina mina, float y);
+void setMinaY(Mina mina, double y);
 
 /** Seta a cor da borda da mina */
 void setMinaCorb(Mina mina, char* cor);
@@ -35,10 +33,10 @@ void setMinaAncora(Mina mina, char* ancora);
 int getMinaId(Mina mina);
 
 /** Retorna a coordenada x da ancora da mina */
-float getMinaX(Mina mina);
+double getMinaX(Mina mina);
 
 /** Retorna a coordenada y da ancora da mina */
-float getMinaY(Mina mina);
+double getMinaY(Mina mina);
 
 /** Retorna a cor da borda da mina */
 char* getMinaCorB(Mina mina);
